Update GridMovement.cs
I've updated the aStar() algorithm.  It should work now, but I'm unable to properly test it (Objects detach their scripts when I download this version to test it; it works in another program I've done).  This script is designed for a world with only one player.  The kinematic arrive algorithm and the aStar information will likely need to be moved to a different script that can be attached to each character when we add multiple characters.